+++ packages/sdk/src/server/types.ts
@@ -1,5 +1,4 @@
 import type { InternalClient, InternalClientFactory } from './internal-client';
-import type { FastifyLoggerInstance } from 'fastify';
 import type { Headers } from '@whatwg-node/fetch';
 import type { GraphQLServerConfig } from './plugins/graphql';
 import type { ConfigurationVariable, WunderGraphConfiguration } from '@wundergraph/protobuf';
@@ -8,6 +7,8 @@ import type { InputVariable } from '../configure/variables';
 import type { ListenOptions, LoggerLevel, ResolvedListenOptions } from '../configure/options';
 import { OperationsClient } from './operations-client';
 import { GraphQLError } from '../client';
+import { TelemetryOptions } from './trace/trace';
+import { RequestLogger } from './logger';
 
 declare module 'fastify' {
 	interface FastifyRequest extends FastifyRequestContext {}
@@ -87,7 +88,7 @@ export interface BaseRequestContext<
 	/**
 	 * The request logger.
 	 */
-	log: FastifyLoggerInstance;
+	log: RequestLogger;
 	/**
 	 * The internal client that is used to communicate with the server.
 	 * @deprecated Use `operations` instead.
@@ -193,6 +194,14 @@ export interface WunderGraphUser<Role extends string = any, CustomClaims extends
 	provider?: string;
 	providerId?: string;
 	userId?: string;
+	/**
+	 * expires indicates the Unix timestamp in milliseconds when the user expires. After that,
+	 * if needs to be revalidated (via the revalidate hook) in order to take effect. If expires
+	 * is <= 0 or undefined, the session lasts forever.
+	 *
+	 * @default undefined
+	 */
+	expires?: number;
 	name?: string;
 	firstName?: string;
 	lastName?: string;
@@ -221,12 +230,21 @@ export interface WunderGraphUser<Role extends string = any, CustomClaims extends
 	rawIdToken?: string;
 }
 
+export interface TracerConfig extends TelemetryOptions {
+	enabled: boolean;
+}
+
 export interface ServerRunOptions {
 	wundergraphDir: string;
 	serverConfig: WunderGraphHooksAndServerConfig<any, any, any, any>;
 	config: WunderGraphConfiguration;
 	gracefulShutdown: boolean;
 	clientFactory: InternalClientFactory;
+}
+
+export interface CreateServerOptions extends ServerRunOptions {
+	serverHost: string;
+	serverPort: number;
 }
 
 /***