+++ packages/testsuite/apps/auth/.wundergraph/wundergraph.server.ts
@@ -32,6 +32,24 @@ export default configureWunderGraphServer(() => ({
 						message: 'denied by mutatingPostAuthentication',
 					};
 				}
+				if (hook.user.userId === 'expiresIn1Second') {
+					return {
+						status: 'ok',
+						user: {
+							...hook.user,
+							expires: Date.now() + 1 * 1000,
+						},
+					};
+				}
+				if (hook.user.userId === 'expiresNow') {
+					return {
+						status: 'ok',
+						user: {
+							...hook.user,
+							expires: Date.now() - 1,
+						},
+					};
+				}
 				const customClaims = hook.user.customClaims ?? {};
 				customClaims.aTestValue = ['anArray'];
 				return {
@@ -48,6 +66,18 @@ export default configureWunderGraphServer(() => ({
 					return {
 						status: 'deny',
 						message: 'denied by revalidate',
+					};
+				}
+				if (hook.user.userId === 'expiresIn1Second') {
+					throw new Error('should not need to revalidate');
+				}
+				if (hook.user.userId === 'expiresNow') {
+					return {
+						status: 'ok',
+						user: {
+							...hook.user,
+							expires: Date.now() + 60 * 1000,
+						},
 					};
 				}
 				return {